project/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
amqp==5.2.0
aniso8601==9.0.1
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.8.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
bcrypt==3.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.6.2
boto==2.49.0
boto3==1.28.0
botocore==1.31.0
Brlapi==0.8.3
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.26.4
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
constantly==23.10.4
contourpy==1.1.0
cron-descriptor==1.4.3
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.9.0
cupshelpers==1.0
cycler==0.11.0
daphne==4.1.0
dbus-python==1.2.18
defender==0.0.3
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.4
django-autoslug==1.9.9
django-celery-beat==2.5.0
django-cleanup==8.1.0
django-cors-headers==3.7.0
django-defender==0.9.7
django-formset-js==0.5.0
django-jet-reboot==1.1.1
django-jquery-js==3.1.1
django-otp==1.3.0
django-ratelimit==3.0.1
django-redis==5.4.0
django-report-builder==6.4.2
django-ses==3.5.2
django-session-timeout==0.1.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
emails==0.6
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fasteners==0.14.1
filelock==3.12.2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.40.0
frappe-bench==5.22.3
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
gitdb==4.0.11
GitPython==3.1.42
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
honcho==1.1.0
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
incremental==22.10.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
jsonify==0.5
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.3.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
metadata-parser==0.12.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
netifaces==0.11.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.0
premailer==3.10.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.11.0
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.1.0
pymacaroons==0.13.0
pyminizip==0.2.6
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.0.9
pypdf==4.0.1
PyPDF2==3.0.1
pypng==0.20220715.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.6
python-dotenv==1.0.0
pytz==2021.1
pyxdg==0.27
pyxirr==0.10.2
PyYAML==5.4.1
qrcode==7.4.2
redis==5.0.1
referencing==0.29.1
reportlab==3.6.6
requests==2.22.0
requests-toolbelt==1.0.0
rpds-py==0.8.4
s3transfer==0.6.1
SecretStorage==3.3.1
SegmentTec-Everything==0.1.0
selenium==4.18.1
semantic-version==2.8.5
service-identity==24.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.18
sqlparse==0.4.1
sympy==1.12
systemd-python==234
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
trio==0.25.0
trio-websocket==0.11.1
triton==2.2.0
twilio==8.13.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.6
whitenoise==6.6.0
wsproto==1.2.0
WTForms==3.1.2
xdg==5
xhtml2pdf==0.2.5
xkit==0.0.0
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.16.2
zope.interface==6.2
